client_agent/pyproject.toml
[project]
name = "github-client-agent"
version = "0.1.0"
description = "Gemini Enterprise-facing client agent that delegates GitHub work to a remote A2A specialist."
authors = [
    { name = "Google Hackathon Team", email = "hackathon@example.com" },
]
dependencies = [
    "google-adk[a2a]>=1.18.0,<2.0.0",
    "google-cloud-aiplatform[evaluation,agent-engines]>=1.130.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "opentelemetry-instrumentation-google-genai>=0.1.0,<1.0.0",
    "gcsfs>=2024.11.0",
    "pyarrow<24",
    "protobuf>=6.31.1,<7.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
]
requires-python = ">=3.10,<3.14"

[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6,<1.0.0",
    "ty>=0.0.1a0",
    "codespell>=2.2.0,<3.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "C", "B", "UP", "RUF"]
ignore = ["E501", "C901", "B006"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ty.environment]
python-version = "3.10"

[tool.ty.src]
exclude = [".venv/**"]

[tool.ty.rules]
unresolved-import = "ignore"
unresolved-attribute = "ignore"
invalid-argument-type = "ignore"
invalid-assignment = "ignore"
invalid-return-type = "ignore"
possibly-missing-attribute = "ignore"
not-subscriptable = "ignore"
deprecated = "ignore"

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,**/*.ipynb"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.agent-starter-pack]
name = "github-client-agent"
description = "Gemini Enterprise-facing client agent that delegates GitHub work to a remote A2A specialist."
base_template = "adk"
agent_directory = "app"
generated_at = "2026-04-21T00:00:00Z"
asp_version = "0.41.1"

[tool.agent-starter-pack.create_params]
deployment_target = "agent_engine"
session_type = "none"
cicd_runner = "skip"
include_data_ingestion = false
datastore = "none"
